move tests to tests directory
