+++ .buffalo.dev.yml
@@ -15,7 +15,7 @@ included_extensions:
 - .env
 build_path: tmp
 build_delay: 200ns
-binary_name: acaturn-Api-build
+binary_name: acaturn-api-build
 command_flags: []
 enable_colors: true
 log_name: buffalo
