+++ actions/routes.go
@@ -1,12 +1,12 @@
 package actions
 
 import (
-	"acaturn_api/actions/Api/V1"
+	"acaturn_api/actions/api/version1"
 	"github.com/gobuffalo/buffalo"
 )
 
 func mountRoutes(app *buffalo.App) {
 	var api = app.Group("/api")
-	var v1 = api.Group("/v1")
-	v1.GET("/", V1.HomeHandler)
+	var v1 = api.Group("/version1")
+	v1.GET("/", version1.HomeHandler)
 }
