+++ database.yml
@@ -8,7 +8,17 @@ development:
   pool: 5
 
 test:
-  url: {{envOr "TEST_DATABASE_URL" "postgres://acaturnuser:acaturn@127.0.0.1:5432/acaturn_api_test?sslmode=disable"}}
+  dialect: postgres
+  database: acaturn_api_test
+  user: {{ envOr "DATABASE_USER" "acaturnuser" }}
+  password: {{ envOr "DATABASE_PASSWORD" "acaturn" }}
+  host: 127.0.0.1
+  pool: 5
 
 production:
-  url: {{envOr "DATABASE_URL" "postgres://acaturnuser:acaturn@127.0.0.1:5432/acaturn_api_production?sslmode=disable"}}
+  dialect: postgres
+  database: acaturn_api_production
+  user: {{ envOr "DATABASE_USER" "acaturnuser" }}
+  password: {{ envOr "DATABASE_PASSWORD" "acaturn" }}
+  host: 127.0.0.1
+  pool: 5
