+++ tests/api_tests/version1_tests/home_test.go
@@ -1,7 +1,9 @@
-package V1
+package version1_tests
+
+var root = "/api/v1"
 
 func (as *V1Suite) TestHomeHandler() {
-	res := as.JSON("/").Get()
+	res := as.JSON(root).Get()
 
 	as.Equal(200, res.Code)
 	as.Contains(res.Body.String(), "Welcome to Buffalo")
